+++ app/services/monitoring/metrics/insights-metrics.service.ts
@@ -4,7 +4,7 @@ import { Observable, of } from "rxjs";
 
 import { Metric, MetricValue, MonitoringMetricList } from "app/models/monitoring";
 import { ArmHttpService } from "app/services/arm-http.service";
-import { BatchAccountService } from "app/services/batch-account.service";
+import { BatchAccountService } from "app/services/batch-account";
 import { ThemeService } from "app/services/themes";
 import { flatMap, map, share } from "rxjs/operators";
 import { CoreCountMetrics } from "./core-count-metrics";
