+++ app/services/ms-graph/core/aad-graph-http.service.ts
@@ -6,7 +6,7 @@ import { Observable, throwError } from "rxjs";
 import { HttpRequestOptions, HttpService, ServerError } from "@batch-flask/core";
 import { UrlUtils } from "@batch-flask/utils";
 import { AdalService } from "app/services/adal";
-import { BatchAccountService } from "app/services/batch-account.service";
+import { BatchAccountService } from "app/services/batch-account";
 import { BatchExplorerService } from "app/services/batch-labs.service";
 import { AADUser } from "client/core/aad/adal/aad-user";
 import { Constants } from "common";
