+++ app/services/ms-graph/core/ms-graph-http.service.ts
@@ -2,7 +2,7 @@ import { HttpHandler } from "@angular/common/http";
 import { Injectable } from "@angular/core";
 import { HttpService, ServerError } from "@batch-flask/core";
 import { AdalService } from "app/services/adal";
-import { BatchAccountService } from "app/services/batch-account.service";
+import { BatchAccountService } from "app/services/batch-account";
 import { BatchExplorerService } from "app/services/batch-labs.service";
 import { AADUser } from "client/core/aad/adal/aad-user";
 import { Observable, throwError } from "rxjs";
